requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
antiorm==1.2.1
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
attrs==22.1.0
autopep8==1.6.0
Babel==2.9.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==7.1.2
colorama==0.4.5
cryptography==37.0.4
db==0.1.1
emoji==2.0.0
frozenlist==1.3.1
google-api-core==2.8.2
google-api-python-client==2.48.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
Google-Images-Search==1.4.3
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.3
joblib==1.1.0
multidict==6.0.2
numpy==1.21.6
outcome==1.2.0
pandas==1.3.5
Pillow==9.2.0
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.0
pycparser==2.21
pyfiglet==0.8.post1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-resize-image==1.1.20
pytz==2022.1
requests==2.28.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
selenium==4.3.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
wsproto==1.1.0
yarl==1.8.1
